fix kernel invocation containing tuple types
CUDA 4.0 changed the way kernel parameters are passed, attempting to match the given parameters with the real kernel arguments. This doesn't work in our setup, since we use struct types to cheat in code generation so skeletons are parametric over dimensionality and array type. I new version of the bindings package provides the necessary behaviour again.